+++ src/main/java/com/mysite/sbb/question/Question.java
@@ -1,29 +1,36 @@
-package com.mysite.sbb;
+package com.mysite.sbb.question;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
+import com.mysite.sbb.answer.Answer;
+
+import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
 @Entity
-public class Answer {
+public class Question {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Integer id;
+
+	@Column(length = 200)
+	private String subject;
 
 	@Column(columnDefinition = "TEXT")
 	private String content;
 
 	private LocalDateTime createDate;
 
-	@ManyToOne
-	private Question question;
+	@OneToMany(mappedBy = "question", cascade = CascadeType.REMOVE)
+	private List<Answer> answerList;
 }
